Fix: USD: Out of bounds access while building armature deform groups
If a USD file has joint indices outside the range of the joints list,
it's possible for our code to assert or crash depending on build
configuration.

One particular file had 289000 indices, nearly all of which were outside
the list which contained just 1 joint value. Instead of continuing when
this is detected, trace an error and immediately return as it's probably
unsafe to continue. This also resulted in many thousands of warning
traces before.

Pull Request: https://projects.blender.org/blender/blender/pulls/133419