Refactor: Anim, rename "unbound slot" to "untyped slot"
An Action slot usually has a specific ID type (`ActionSlot::idtype`).
When this is set to 0, it is not specifc to any particular ID type. The
latter case was called "unbound", and is now called "untyped".

Note that untyped slots are given an ID type as soon as they are
actually assigned to any ID.

No functional changes.

Pull Request: https://projects.blender.org/blender/blender/pulls/133651