Anim: allow reordering Slots in Action Editor channel list
This updates the Move Channels operator in the Action Editor channel list to
also work with Slots, allowing Slots to be reordered within an Action.

Note that the Move Channels operator already had interesting behavior when
channels are selected in multiple levels of the channel hierarchy (e.g. when
both channel groups and f-curve channels are selected).  For Slots, I've chosen
a simpler behavior: if any Slot is moved, all non-Slot channels (i.e. those
lower in the hierarchy) are left alone.

In the future we may want to change channel group rearrangement to work the same
way with respect to f-curve channels, but I've left that alone for now.

Pull Request: https://projects.blender.org/blender/blender/pulls/133328